show fund status on providers overview page
