+++ react/src/dashboard/components/pages/organizations-funds/OrganizationFunds.tsx
@@ -28,6 +28,7 @@ import ModalFundTopUp from '../../modals/ModalFundTopUp';
 import useTranslate from '../../../hooks/useTranslate';
 import useAssetUrl from '../../../hooks/useAssetUrl';
 import TableEmptyValue from '../../elements/table-empty-value/TableEmptyValue';
+import FundStateLabels from '../../elements/resource-states/FundStateLabels';
 
 export default function OrganizationFunds() {
     const translate = useTranslate();
@@ -56,12 +57,6 @@ export default function OrganizationFunds() {
         { key: 'paused', name: translate(`components.organization_funds.states.paused`) },
         { key: 'closed', name: translate(`components.organization_funds.states.closed`) },
     ]);
-
-    const [stateLabels] = useState({
-        active: 'label-success',
-        paused: 'label-warning',
-        closed: 'label-default',
-    });
 
     const [{ funds_type }, setQueryParams] = useQueryParams(
         { funds_type: withDefault(StringParam, 'active') },
@@ -432,21 +427,7 @@ export default function OrganizationFunds() {
                                             )}
 
                                             <td>
-                                                {!fund.archived && stateLabels[fund.state] ? (
-                                                    <span className={`label ${stateLabels[fund.state] || ''}`}>
-                                                        {translate(
-                                                            `components.organization_funds.states.${fund.state}`,
-                                                        )}
-                                                    </span>
-                                                ) : (
-                                                    <TableEmptyValue />
-                                                )}
-
-                                                {fund.archived && (
-                                                    <span className="label label-default">
-                                                        {translate('components.organization_funds.states.archived')}
-                                                    </span>
-                                                )}
+                                                <FundStateLabels fund={fund} />
                                             </td>
 
                                             <td className="td-narrow text-right">
